Support re-using FileStorages in subsequent test runs.
This was an experiment to see if I could make the inner development loop
faster by using FileStorage instead of DemoStorage for ZODB changes in
layers and then re-using the same FileStorage on subsequent test runs.
I had hoped that this would reduce the time spent waiting for layer
setup when running just a few tests to check the effect of your work.

Having prototyped it and timed it, it does cut layer set up time by
close to half, but that turns out not to make that much of an overall
difference unless the layer has a *much* longer set up time.  Since the
changes to enable this are not insignificant and running full suites
showed various isolation errors, I'm calling this a failure.